tests/wpt/tests/webmessaging/with-ports/005.html

<!doctype html>
<title>resolving a same origin targetOrigin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  postMessage('', location.protocol + '//' + location.host, []);
  onmessage = this.step_func(function(e) {
    assert_equals(e.data, '');
    this.done();
  });
});
</script>
